move cmd/redpanda to its own package
