+++ src/go/rpk/pkg/cli/cmd/root_darwin.go
@@ -10,10 +10,11 @@
 package cmd
 
 import (
+	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/cmd/redpanda"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 )
 
 func addPlatformDependentCmds(fs afero.Fs, cmd *cobra.Command) {
-	cmd.AddCommand(NewRedpandaDarwinCommand(fs))
+	cmd.AddCommand(redpanda.NewRedpandaDarwinCommand(fs))
 }
