+++ src/go/rpk/pkg/cli/cmd/root_linux.go
@@ -11,20 +11,21 @@ package cmd
 
 import (
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/cmd/iotune"
-	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/redpanda"
+	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/cmd/redpanda"
+	rp "github.com/redpanda-data/redpanda/src/go/rpk/pkg/redpanda"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 )
 
 func addPlatformDependentCmds(fs afero.Fs, cmd *cobra.Command) {
-	cmd.AddCommand(NewRedpandaCommand(fs, redpanda.NewLauncher()))
+	cmd.AddCommand(redpanda.NewCommand(fs, rp.NewLauncher()))
 	cmd.AddCommand(iotune.NewCommand(fs))
 
 	// deprecated
 	cmd.AddCommand(NewCheckCommand(fs))
 	cmd.AddCommand(NewConfigCommand(fs))
 	cmd.AddCommand(NewModeCommand(fs))
-	cmd.AddCommand(NewStartCommand(fs, redpanda.NewLauncher()))
+	cmd.AddCommand(NewStartCommand(fs, rp.NewLauncher()))
 	cmd.AddCommand(NewStatusCommand(fs))
 	cmd.AddCommand(NewStopCommand(fs))
 	cmd.AddCommand(NewTuneCommand(fs))
